pyproject.toml
[tool.poetry]
name = "allergy-checker"
version = "0.1.0"
description = ""
authors = ["Anurag Singh <anurag.singh@macmillan.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.25.1"
flask = "^3.0.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
